+++ src/form/b-button/b-button.ts
@@ -32,6 +32,7 @@ import iData, {
 
 	component,
 	prop,
+	system,
 	computed,
 	wait,
 	p,
@@ -283,6 +284,13 @@ class bButton extends iData implements iAccess, iOpenToggle, iVisible, iWidth, i
 	@prop(String)
 	readonly dropdown: string = 'bottom';
 
+	/**
+	 * Additional attributes are provided to an "internal" (native) button tag
+	 * @see [[bButton.$refs.button]]
+	 */
+	@prop({type: Object, required: false})
+	readonly attrsProp?: Dictionary;
+
 	/** @see [[iAccess.isFocused]] */
 	@computed({dependencies: ['mods.focused']})
 	get isFocused(): boolean {
@@ -304,6 +312,18 @@ class bButton extends iData implements iAccess, iOpenToggle, iVisible, iWidth, i
 		return this.$refs.file?.files ?? undefined;
 	}
 
+	/**
+	 * True if the component has a dropdown area
+	 */
+	get hasDropdown(): boolean {
+		return Boolean(
+			this.vdom.getSlot('dropdown') && (
+				this.isFunctional ||
+				this.opt.ifOnce('opened', this.m.opened !== 'false') > 0 && delete this.watchModsStore.opened
+			)
+		);
+	}
+
 	/** @inheritDoc */
 	static readonly mods: ModsDecl = {
 		...iAccess.mods,
@@ -322,10 +342,22 @@ class bButton extends iData implements iAccess, iOpenToggle, iVisible, iWidth, i
 		]
 	};
 
+	/**
+	 * Additional attributes that are provided to an "internal" (native) button tag
+	 * @see [[bButton.attrsProp]]
+	 */
+	@system<bButton>({
+		after: 'mods',
+		init: (o) => o.sync.link(o.normalizeAttrs.bind(o))
+	})
+
+	protected attrs?: Dictionary;
+
 	/** @override */
 	protected readonly $refs!: {
 		button: HTMLButtonElement;
 		file?: HTMLInputElement;
+		dropdown?: Element;
 	};
 
 	/**
@@ -341,6 +373,31 @@ class bButton extends iData implements iAccess, iOpenToggle, iVisible, iWidth, i
 		}
 	}
 
+	/**
+	 * Normalizes the specified additional attributes and returns it
+	 *
+	 * @see [[bButton.attrs]]
+	 * @param [attrs]
+	 */
+	protected normalizeAttrs(attrs: Dictionary = {}): Dictionary {
+		attrs = {...attrs};
+
+		if (this.type === 'link') {
+			attrs.href = this.href;
+
+		} else {
+			attrs.type = this.type;
+			attrs.form = this.form;
+		}
+
+		if (this.hasDropdown) {
+			attrs['aria-controls'] = this.dom.getId('dropdown');
+			attrs['aria-expanded'] = this.mods.opened;
+		}
+
+		return attrs;
+	}
+
 	/** @see [[iOpenToggle.initCloseHelpers]] */
 	@p({hook: 'beforeDataCreate', replace: false})
 	protected initCloseHelpers(events?: CloseHelperEvents): void {
@@ -348,7 +405,16 @@ class bButton extends iData implements iAccess, iOpenToggle, iVisible, iWidth, i
 	}
 
 	/** @override */
+	protected initBaseAPI(): void {
+		super.initBaseAPI();
+		this.normalizeAttrs = this.instance.normalizeAttrs.bind(this);
+	}
+
+	/** @override */
 	protected initModEvents(): void {
+		const
+			{localEmitter: $e} = this;
+
 		super.initModEvents();
 
 		iProgress.initModEvents(this);
@@ -356,7 +422,12 @@ class bButton extends iData implements iAccess, iOpenToggle, iVisible, iWidth, i
 		iOpenToggle.initModEvents(this);
 		iVisible.initModEvents(this);
 
-		this.localEmitter.on('block.mod.*.disabled.*', (e: ModEvent) => this.waitStatus('ready', () => {
+		$e.on('block.mod.*.opened.*', (e: ModEvent) => this.waitStatus('ready', () => {
+			const expanded = e.value !== 'false' && e.type !== 'remove';
+			this.$refs.button.setAttribute('aria-expanded', String(expanded));
+		}));
+
+		$e.on('block.mod.*.disabled.*', (e: ModEvent) => this.waitStatus('ready', () => {
 			const {
 				button,
 				file
@@ -370,7 +441,7 @@ class bButton extends iData implements iAccess, iOpenToggle, iVisible, iWidth, i
 			}
 		}));
 
-		this.localEmitter.on('block.mod.*.focused.*', (e: ModEvent) => this.waitStatus('ready', () => {
+		$e.on('block.mod.*.focused.*', (e: ModEvent) => this.waitStatus('ready', () => {
 			const
 				{button} = this.$refs;